Clean up - made one class handle the logic for very basic appbars
