+++ app/lib/pages/changelog_page.dart
@@ -1,12 +1,10 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart' show rootBundle;
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:localsend_app/gen/assets.gen.dart';
 import 'package:localsend_app/gen/strings.g.dart';
 import 'package:localsend_app/util/ui/nav_bar_padding.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 
 class ChangelogPage extends StatelessWidget {
   const ChangelogPage();
@@ -14,11 +12,7 @@ class ChangelogPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac(t.changelogPage.title)
-          : AppBar(
-              title: Text(t.changelogPage.title),
-            ),
+      appBar: basicLocalSendAppbar(t.changelogPage.title),
       body: FutureBuilder(
         future: rootBundle.loadString(Assets.changelog), // ignore: discarded_futures
         builder: (context, data) {
