+++ app/lib/pages/debug/debug_page.dart
@@ -8,7 +8,7 @@ import 'package:localsend_app/pages/debug/security_debug_page.dart';
 import 'package:localsend_app/provider/app_arguments_provider.dart';
 import 'package:localsend_app/provider/persistence_provider.dart';
 import 'package:localsend_app/util/shared_preferences/shared_preferences_file.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/debug_entry.dart';
 import 'package:refena_flutter/refena_flutter.dart';
 import 'package:routerino/routerino.dart';
@@ -24,11 +24,7 @@ class DebugPage extends StatelessWidget {
     final store = SharedPreferencesStorePlatform.instance;
 
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac('Debugging')
-          : AppBar(
-              title: Text('Debugging'),
-            ),
+      appBar: basicLocalSendAppbar('Debugging'),
       body: ListView(
         padding: const EdgeInsets.only(left: 20, right: 20, top: 10, bottom: 30),
         children: [
