+++ app/lib/pages/debug/http_logs_page.dart
@@ -1,10 +1,8 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:localsend_app/provider/logging/http_logs_provider.dart';
 import 'package:localsend_app/widget/copyable_text.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:refena_flutter/refena_flutter.dart';
 
@@ -17,11 +15,7 @@ class HttpLogsPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final logs = context.ref.watch(httpLogsProvider);
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac('HTTP Logs')
-          : AppBar(
-              title: Text('HTTP Logs'),
-            ),
+      appBar: basicLocalSendAppbar('HTTP Logs'),
       body: ResponsiveListView(
         padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 30),
         children: [
