+++ app/lib/pages/debug/security_debug_page.dart
@@ -1,8 +1,6 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:localsend_app/provider/security_provider.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/debug_entry.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:refena_flutter/refena_flutter.dart';
@@ -14,11 +12,7 @@ class SecurityDebugPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final securityContext = context.ref.watch(securityProvider);
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac('Security Debugging')
-          : AppBar(
-              title: Text('Security Debugging'),
-            ),
+      appBar: basicLocalSendAppbar('Security Debugging'),
       body: ResponsiveListView(
         padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 30),
         maxWidth: 700,
