+++ app/lib/pages/language_page.dart
@@ -1,9 +1,7 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:localsend_app/gen/strings.g.dart';
 import 'package:localsend_app/provider/settings_provider.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:refena_flutter/refena_flutter.dart';
 
@@ -30,11 +28,7 @@ class _LanguagePageState extends State<LanguagePage> {
     final t = Translations.of(context);
     final activeLocale = context.ref.watch(settingsProvider.select((s) => s.locale));
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac(t.sendTab.selection.title)
-          : AppBar(
-              title: Text(t.sendTab.selection.title),
-            ),
+      appBar: basicLocalSendAppbar(t.sendTab.selection.title),
       body: ResponsiveListView(
         padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 20),
         children: [
