+++ app/lib/pages/send_page.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:common/model/device.dart';
 import 'package:common/model/session_status.dart';
@@ -13,7 +12,7 @@ import 'package:localsend_app/util/favorites.dart';
 import 'package:localsend_app/util/native/taskbar_helper.dart';
 import 'package:localsend_app/widget/animations/initial_fade_transition.dart';
 import 'package:localsend_app/widget/animations/initial_slide_transition.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/dialogs/error_dialog.dart';
 import 'package:localsend_app/widget/list_tile/device_list_tile.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
@@ -88,11 +87,7 @@ class _SendPageState extends State<SendPage> with Refena {
       },
       canPop: true,
       child: Scaffold(
-        appBar: widget.showAppBar
-            ? Platform.isMacOS
-                ? appBarMac('')
-                : AppBar()
-            : null,
+        appBar: widget.showAppBar ? basicLocalSendAppbar('') : null,
         body: SafeArea(
           child: Center(
             child: ConstrainedBox(
