+++ app/lib/pages/tabs/receive_tab.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:flutter/material.dart';
 import 'package:localsend_app/gen/strings.g.dart';
@@ -9,6 +7,7 @@ import 'package:localsend_app/pages/receive_history_page.dart';
 import 'package:localsend_app/pages/tabs/receive_tab_vm.dart';
 import 'package:localsend_app/provider/animation_provider.dart';
 import 'package:localsend_app/util/ip_helper.dart';
+import 'package:localsend_app/util/native/platform_check.dart';
 import 'package:localsend_app/widget/animations/initial_fade_transition.dart';
 import 'package:localsend_app/widget/column_list_view.dart';
 import 'package:localsend_app/widget/custom_icon_button.dart';
@@ -33,7 +32,7 @@ class ReceiveTab extends StatelessWidget {
 
     return Stack(
       children: [
-        Platform.isMacOS
+        checkPlatform([TargetPlatform.macOS])
             ? SizedBox(height: 50, child: MoveWindow())
             : SizedBox(height: 0, width: 0), // makes the top part that's not occupied by another widget draggable
         Center(
