+++ app/lib/pages/troubleshoot_page.dart
@@ -6,8 +6,8 @@ import 'package:localsend_app/gen/strings.g.dart';
 import 'package:localsend_app/provider/settings_provider.dart';
 import 'package:localsend_app/util/native/cmd_helper.dart';
 import 'package:localsend_app/util/native/platform_check.dart';
+import 'package:localsend_app/widget/custom_basic_appbar.dart';
 import 'package:localsend_app/widget/custom_icon_button.dart';
-import 'package:localsend_app/widget/custom_macos_appbar.dart';
 import 'package:localsend_app/widget/dialogs/not_available_on_platform_dialog.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:refena_flutter/refena_flutter.dart';
@@ -19,11 +19,7 @@ class TroubleshootPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final settings = context.ref.watch(settingsProvider);
     return Scaffold(
-      appBar: Platform.isMacOS
-          ? appBarMac(t.troubleshootPage.title)
-          : AppBar(
-              title: Text(t.troubleshootPage.title),
-            ),
+      appBar: basicLocalSendAppbar(t.troubleshootPage.title),
       body: ResponsiveListView(
         padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 30),
         children: [
