+++ app/lib/util/native/tray_helper.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:flutter/foundation.dart';
 import 'package:localsend_app/gen/assets.gen.dart';
 import 'package:localsend_app/gen/strings.g.dart';
@@ -81,6 +82,7 @@ Future<void> showFromTray() async {
     // This will crash on Windows
     // https://github.com/localsend/localsend/issues/32
     await windowManager.setSkipTaskbar(false);
+    appWindow.show();
   }
 
   // Enable animations
